test(shell): move boundary coverage into existing files
